% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.stabel.R
\name{cv.stabel}
\alias{cv.stabel}
\title{Cross-validation for stabel}
\usage{
cv.stabel(
  X,
  Y,
  family = "binomial",
  fit.fun = c("Lasso", "sparseSVM"),
  eval.metric.Lasso = NULL,
  eval.metric.sparseSVM = NULL,
  nfolds = 10,
  seed = NULL
)
}
\arguments{
\item{X}{Input matrix.}

\item{Y}{Outcome vector.}

\item{family}{A description of the error distribution and link function to be used in the model. Even though Lasso works for both \code{"gaussian"} and \code{"binomial"} family, \code{sparseSVM} works only for \code{"binomial"} family. Default is \code{"binomial"}.}

\item{fit.fun}{List the name of the planned variable selection methods. Default is \code{c("Lasso", "sparseSVM")}.  It can be changed to either only \code{"Lasso"} or \code{"sparseSVM"}.}

\item{eval.metric.Lasso}{Loss to use for cross-validation. Currently five options, not all available for all models. Check \code{glmnet} package for more details. If \code{fit.fun} includes \code{"Lasso"}, then \code{eval.metric.Lasso} must be specified,  otherwise it will throw an error.}

\item{eval.metric.sparseSVM}{The metric used to choose optimal lambda. Current version only supports "me": misclassification error. If fit.fun includes \code{"sparseSVM"}, then \code{eval.metric.sparseSVM} must be specified,  otherwise it will throw an error.}

\item{nfolds}{Number of folds for cross-validation. Default is 10.}

\item{seed}{A random seed for reproducibility of results.}
}
\value{
A list containing the following components:
\item{bestlam.Lasso}{The optimal regularization parameter (\eqn{\lambda}) for the Lasso model selected through cross-validation.}
\item{bestlam.sparseSVM}{The optimal regularization parameter (\eqn{\lambda}) for the sparseSVM model selected through cross-validation.}
\item{family}{Same as above.}
\item{eval.metric.Lasso}{The evaluation metric used to determine the optimal model for Lasso.}
\item{eval.metric.sparseSVM}{The evaluation metric used to determine the optimal model for sparseSVM.}
\item{nfolds}{Same as above.}
}
\description{
The \code{cv.stabel} function performs cross-validation for stabel, allowing the use of either Lasso, sparseSVM, or both for variable selection and model evaluation.
}
\examples{
cv.stabel()
}
\keyword{stabel}
