% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stabel.pred.R
\name{stabel.pred}
\alias{stabel.pred}
\title{Prediction using stabel}
\usage{
stabel.pred(
  X,
  Y,
  newX = NULL,
  newY = NULL,
  method = "method.NNLS",
  SL.library,
  family = "binomial",
  nfolds = 10,
  thr.prob = 0.5,
  use.youden = FALSE,
  target.specificity = NULL,
  target.sensitivity = NULL,
  seed = NULL
)
}
\arguments{
\item{X}{The input matrix of predictors (features) used to train the model.}

\item{Y}{The outcome vector corresponding to the input matrix.}

\item{newX}{A matrix of new samples. If provided, predictions will be made for these new observations. Should have the same column names as \code{X}.}

\item{newY}{The true outcomes for \code{newX}, used for evaluating the model's performance on test data. If provided, metrics like AUC or MSE can be calculated by comparing predictions to \code{newY}.}

\item{method}{The method used for combining models in the ensemble. For example, the default method \code{"method.NNLS"} indicates non-negative least squares. Can be changed to other ensemble methods supported by the \code{SuperLearner} package.}

\item{SL.library}{A list of algorithms/models to be included in the ensemble learning. Check the \code{SuperLearner} package for more details.}

\item{family}{Currently allows \code{"gaussian"} or \code{"binomial"} to describe the error distribution. Link function information will be ignored and should be contained in the method argument below. Default is \code{"binomial"}.}

\item{nfolds}{Number of folds for cross-validation. Default is 10.}

\item{thr.prob}{The threshold probability for classifying predictions into binary classes. Default is 0.5. It should be set to \code{NULL} if family is not binomial.}

\item{use.youden}{Youden index as threshold value or cutoff point for classifying predictions into binary classes. Default is \code{TRUE}. Both \code{thr.prob} and \code{use.youden} cannot be used together.}

\item{target.specificity}{The desired specificity at which sensitivity will be evaluated. It is optional.}

\item{target.sensitivity}{The desired sensitivity at which specificity will be evaluated. It is optional.}

\item{seed}{A random seed for reproducibility of results.}
}
\value{
A list containing the following components:
\item{method}{Same as above.}
\item{family}{Same as above.}
\item{nfolds}{Same as above.}
\item{cvRisk}{ Cross-validated risk for each candidate algorithm in the SuperLearner ensemble.}
\item{coef}{Calculated weight given to each candidate algorithm based on cvRisk.}
\item{roc.object}{ROC object used for plotting the ROC curve or computing confidence intervals of the AUC (available for \code{family = "binomial"}).}
\item{auc}{Area under the ROC curve (AUC) for classification models (available for \code{family = "binomial"}).}
\item{predicted.probs}{Predicted probabilities for binary classification (available for \code{family = "binomial"}).}
\item{predicted.responses}{Predicted responses for continuous outcomes (available for \code{family = "gaussian"}).}
\item{threshold.probability}{The threshold probability used for classification (available for \code{family = "binomial"}).}
\item{confusion.matrix}{Confusion matrix summarizing the classification results (available for \code{family = "binomial"}).}
\item{predition.accuracy}{The overall accuracy of the predictions.}
\item{sensitivity}{Sensitivity (true positive rate) of the predictions.}
\item{specificity}{Specificity (true negative rate) of the predictions.}
\item{precision}{Precision (positive predictive value) of the predictions.}
\item{negative.predictive.value}{Negative predictive value of the predictions.}
\item{positive.predictive.value}{Positive predictive value of the predictions.}
\item{false.positve.rate}{False positive rate of the predictions.}
\item{false.negative.rate}{False negative rate of the predictions.}
\item{F1.score}{F1 score, the harmonic mean of precision and sensitivity.}
\item{target.specificity}{User-specified target specificity.}
\item{target.sensitivity}{User-specified target sensitivity.}
\item{sensitivity.at.specificity}{Sensitivity calculated at the specified target specificity.}
\item{specificity.at.sensitivity}{ Specificity calculated at the specified target sensitivity.}
\item{mse}{Mean squared error of the predictions (available for \code{family = "gaussian"}).}
\item{rmse}{Root mean squared error of the predictions (available for \code{family = "gaussian"}).}
\item{mae}{Mean absolute error of the predictions (available for \code{family = "gaussian"}).}
}
\description{
The \code{stabel.pred} function implements prediction using ensemble learning by combining multiple predictive models to improve overall accuracy and robustness.
}
\examples{
stabel.pred()
}
\keyword{stabel}
