% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stabel.sim.R
\name{stabel.sim}
\alias{stabel.sim}
\title{Simulating data for stabel}
\usage{
stabel.sim(
  n,
  p,
  beta,
  family = "binomial",
  mu = rep(0, p),
  rho = NULL,
  corstr = "Independent",
  custom.corr = NULL,
  seed = NULL
)
}
\arguments{
\item{n}{The total number of observations to be generated in the dataset.}

\item{p}{The number of covariates (or predictors) in the dataset.}

\item{beta}{A vector of coefficients for the covariates, representing the true relationship between the predictors and the outcome variable. The length should be same as p.}

\item{family}{Specifies the type of outcome variable to be generated. Options include \code{"gaussian"} for continuous outcomes and \code{"binary"} for binary outcomes. Default is \code{"binomial"}.}

\item{mu}{A vector of means for the covariates. It specifies the center of the multivariate normal distribution used for generating covariates. Default is a vector of zero.}

\item{rho}{The correlation parameter that controls strength of the relationship between covariates. Default is \code{NULL} as the default \code{corstr} is \code{"Independent"}.}

\item{corstr}{The correlation structure to use for generating the covariates. Options include: \code{"Independent"}, \code{"Exchangeable"}, and \code{"AR1"}. Default is \code{"Independent"}.}

\item{custom.corr}{A user-specified correlation matrix with dimension \code{p x p}. Default is \code{NULL}.}

\item{seed}{A random seed for reproducibility of results.}
}
\value{
A list containing the following components:
\item{X}{A matrix of size \code{n x p} containing the simulated covariates.}
\item{Y}{A vector of length \code{n} containing the simulated outcomes. Continuous outcomes for \code{family = "gaussian"} and binary outcomes (0 or 1) for \code{family = "binary"}.}
}
\description{
The \code{stabel.sim} function generates a simulated dataset with multivariate normal covariates and corresponding outcomes based on specified model parameters. The simulation supports both continuous and binary outcomes and allows for flexible correlation structures.
}
\examples{
stabel.sim()

}
\keyword{stabel}
